fix: helm_args plugin discover find fix (#369)
* fix: helm_args plugin find fix

The plugin shouldn't look for env specific values files in the chart directory. Instead it should look for the default `values.yaml` file. The other parts of the plugin discovery condition is file.

* fix: change the plugin order for discovery

Moved avp-helm-kustomize up so it will be used instead of avp-help-args when kustomization file is also present.

* fix: helm_args plugin discover when kustomization is not present

A new condition is added so plugin is only triggered when kustomization file is not present.