chore: fix api resource version
For whatever reason the automatic determination of api-resource version for podDisruptionBudget fails. The Helm Chart provides a way to set apiVersion for that.

ArgoCD error was:

```
The Kubernetes API could not find version "v1beta1" of policy/PodDisruptionBudget for requested resource kyverno/kyverno-background-controller. Version "v1" of policy/PodDisruptionBudget is installed on the destination cluster.
``